Rename exception.
