+++ service/src/main/java/bio/terra/tanagra/vumc/admin/service/authentication/HttpUtils.java
@@ -1,8 +1,9 @@
 package bio.terra.tanagra.vumc.admin.service.authentication;
 
-import bio.terra.tanagra.vumc.admin.service.authentication.exception.SystemException;
 import java.time.Duration;
 import java.util.function.Predicate;
+
+import bio.terra.common.exception.InternalServerErrorException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -24,7 +25,7 @@ public class HttpUtils {
    * @param makeRequest function to perform the request
    * @param isRetryable function to test whether the exception is retryable or not
    * @throws E if makeRequest throws an exception that is not retryable
-   * @throws SystemException if the maximum number of retries is exhausted, and the last attempt
+   * @throws InternalServerErrorException if the maximum number of retries is exhausted, and the last attempt
    *     threw a retryable exception
    */
   public static <E extends Exception> void callWithRetries(
@@ -50,7 +51,7 @@ public class HttpUtils {
    * @param <T> type of the response object (i.e. return type of the makeRequest function)
    * @return the response object
    * @throws E if makeRequest throws an exception that is not retryable
-   * @throws SystemException if the maximum number of retries is exhausted, and the last attempt
+   * @throws InternalServerErrorException if the maximum number of retries is exhausted, and the last attempt
    *     threw a retryable exception
    */
   public static <T, E extends Exception> T callWithRetries(
@@ -70,7 +71,7 @@ public class HttpUtils {
    * @param sleepDuration time to sleep between tries
    * @return the response object
    * @throws E if makeRequest throws an exception that is not retryable
-   * @throws SystemException if the maximum number of retries is exhausted, and the last attempt
+   * @throws InternalServerErrorException if the maximum number of retries is exhausted, and the last attempt
    *     threw a retryable exception
    */
   public static <T, E extends Exception> T callWithRetries(
@@ -92,7 +93,7 @@ public class HttpUtils {
    * @param <T> type of the response object (i.e. return type of the makeRequest function)
    * @return the response object
    * @throws E if makeRequest throws an exception that is not retryable
-   * @throws SystemException if the maximum number of retries is exhausted, and the last attempt
+   * @throws InternalServerErrorException if the maximum number of retries is exhausted, and the last attempt
    *     threw a retryable exception
    */
   public static <T, E extends Exception> T pollWithRetries(
@@ -119,7 +120,7 @@ public class HttpUtils {
    * the result is returned.
    *
    * <p>- If the last attempt threw a retryable exception, then this method re-throws that last
-   * exception wrapped in a {@link SystemException} with a timeout message.
+   * exception wrapped in a {@link InternalServerErrorException} with a timeout message.
    *
    * @param <T> type of the response object (i.e. return type of the makeRequest function)
    * @param makeRequest function to perform the request
@@ -129,7 +130,7 @@ public class HttpUtils {
    * @param sleepDuration time to sleep between tries
    * @return the response object
    * @throws E if makeRequest throws an exception that is not retryable
-   * @throws SystemException if the maximum number of retries is exhausted, and the last attempt
+   * @throws InternalServerErrorException if the maximum number of retries is exhausted, and the last attempt
    *     threw a retryable exception
    */
   public static <T, E extends Exception> T pollWithRetries(
@@ -177,7 +178,7 @@ public class HttpUtils {
     } while (numTries <= maxCalls);
 
     // request with retries timed out: re-throw the last exception
-    throw new SystemException(
+    throw new InternalServerErrorException(
         "Request with retries timed out after "
             + numTries
             + " tries: "
