+++ service/src/main/java/bio/terra/tanagra/vumc/admin/service/authentication/InvalidCredentialsException.java
@@ -1,0 +1,13 @@
+package bio.terra.tanagra.vumc.admin.service.authentication;
+
+import bio.terra.common.exception.UnauthorizedException;
+
+public class InvalidCredentialsException extends UnauthorizedException {
+  public InvalidCredentialsException(String message) {
+    super(message);
+  }
+
+  public InvalidCredentialsException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}
